refactor: remove duplicate helper functions and upgrade pagination logic
Phase 1 Refactoring - Remove Dead Code and Improve Pagination

CHANGES:
- Remove 104 lines of duplicate dead code from InstantlyApi.node.ts
  - Removed duplicate formatDateForApi function (47 lines, never called)
  - Removed duplicate paginateInstantlyApi function (56 lines, never called)
  - Added proper imports from /functions directory (2 lines)

- Upgrade paginationHelpers.ts with robust pagination logic (+16 lines)
  - Now handles 3 response formats: response.items, response.data, Array
  - Added empty page detection to prevent unnecessary API calls
  - Improved logging for better debugging
  - More resilient pagination with dual checks (hasMore flag + empty items)

IMPACT:
- Net change: -88 lines of code (cleaner codebase)
- InstantlyApi.node.ts: 412 → 307 lines (-105 lines)
- paginationHelpers.ts: 55 → 71 lines (+16 lines)
- Zero TypeScript compilation errors
- All operations continue to work correctly

TESTING COMPLETED:
✅ Phase 2 - Critical Operations (3/3 passed):
  - Analytics → Get Campaign Analytics (formatDateForApi tested)
  - Account → Get Many Accounts (paginateInstantlyApi tested)
  - Campaign → Get Many Campaigns (paginateInstantlyApi tested)

✅ Phase 3 - Smoke Tests (3/3 passed):
  - Account → Get Account (no regressions)
  - Campaign → Get Campaign (no regressions)
  - Lead → Get Lead (no regressions)

✅ Phase 4 - Final Verification:
  - Server starts with 0 errors
  - Pagination handles all response formats
  - Date formatting works with n8n DateTime objects
  - No console errors or warnings
  - Production functionality safe for 8,000+ users

BREAKING CHANGES: None - all changes are internal refactoring

FILES MODIFIED:
- nodes/InstantlyApi/InstantlyApi.node.ts
- nodes/InstantlyApi/functions/paginationHelpers.ts
